refactor: OrderTable, TableGroup 패키지 분리
- OrderTable, TableGroup 둘 다 AbstractAggregateRoot로 되면서 패키지 분리
