+++ src/main/java/kitchenpos/tablegroup/dto/TableGroupResponse.java
@@ -1,11 +1,12 @@
-package kitchenpos.table.dto;
+package kitchenpos.tablegroup.dto;
 
 import static java.util.stream.Collectors.toList;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import kitchenpos.table.domain.OrderTables;
-import kitchenpos.table.domain.TableGroup;
+import kitchenpos.table.dto.OrderTableResponse;
+import kitchenpos.tablegroup.domain.TableGroup;
 
 public class TableGroupResponse {
