+++ src/test/java/kitchenpos/ui/TableGroupAcceptanceTest.java
@@ -10,9 +10,9 @@ import io.restassured.response.Response;
 import java.util.List;
 import kitchenpos.AcceptanceTest;
 import kitchenpos.table.dto.OrderTableResponse;
-import kitchenpos.table.dto.TableGroupResponse;
-import kitchenpos.table.dto.TableGroupsRequest;
-import kitchenpos.table.dto.TableGroupsRequest.OrderTableRequest;
+import kitchenpos.tablegroup.dto.TableGroupResponse;
+import kitchenpos.tablegroup.dto.TableGroupsRequest;
+import kitchenpos.tablegroup.dto.TableGroupsRequest.OrderTableRequest;
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
